Make restructure-namespaces rule configurable
Let users provide a list of namespaces for destructuring.
