+++ lib/rules/destructure-namespaces.js
@@ -5,9 +5,16 @@ var message = "Assign properties from the Ember namespace to variables before " 
               "extract the properties you plan to use: `const { computed, get } = Ember;`";
 
 var rule = function(context) {
+  var protectedNamesOption = context.options[0];
+  var protectedNames = {};
+
+  for (var i = 0; i < protectedNamesOption.length; i++) {
+    protectedNames[protectedNamesOption[i]] = true;
+  }
+
   return {
     'MemberExpression:exit': function(node) {
-      if (node.object.name === 'Ember' &&
+      if (protectedNames.hasOwnProperty(node.object.name) &&
           node.parent.type !== 'VariableDeclarator' &&
           node.parent.type !== 'AssignmentExpression') {
 
@@ -19,5 +26,14 @@ var rule = function(context) {
     }
   };
 };
+
+rule.schema = [
+  {
+      type: "array",
+      items: {
+        type: "string"
+      }
+  }
+];
 
 module.exports = { rule: rule, message: message };
